refactor: 페이지 구성 수정
