+++ carea/lib/app/common/component/category.dart
@@ -1,0 +1,9 @@
+// 서버 연결 전 임시 카테고리 목록
+
+const category = [
+  {'text': "최신글"},
+  {"text": "자유 게시판"},
+  {"text": "경제/금융"},
+  {"text": "생활"},
+  {"text": "진로"}
+];
