+++ carea/lib/app/modules/community/view/community_screen.dart
@@ -1,9 +1,10 @@
 import 'package:carea/app/common/component/board_button.dart';
-import 'package:carea/app/common/component/bottom_bar.dart';
 import 'package:carea/app/common/layout/default_layout.dart';
+import 'package:carea/app/modules/community/view/post_list.dart';
+import 'package:carea/app/modules/user/view/mypage_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:carea/app/common/const/app_colors.dart';
-import 'package:go_router/go_router.dart';
+import 'package:carea/app/common/component/category.dart';
 
 class CommunityScreen extends StatefulWidget {
   const CommunityScreen({super.key});
@@ -29,7 +30,10 @@ class _CommunityScreenState extends State<CommunityScreen> {
               icon: const Icon(Icons.notifications_none_outlined)),
           IconButton(
               onPressed: () {
-                context.go('/mypage');
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(builder: (context) => const MypageScreen()),
+                );
               },
               icon: const Icon(Icons.person_2_outlined))
         ],
@@ -38,47 +42,25 @@ class _CommunityScreenState extends State<CommunityScreen> {
       ),
       body: DefaultLayout(
         child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            const SizedBox(height: 20),
-            BoardButton(
-                onPressed: () {
-                  context.push('/community/newest');
+          mainAxisAlignment: MainAxisAlignment.start,
+          crossAxisAlignment: CrossAxisAlignment.center,
+          children: <Widget>[
+            Flexible(
+              child: ListView.builder(
+                itemCount: category.length,
+                itemBuilder: (BuildContext context, int index) {
+                  return BoardButton(
+                      onPressed: () {
+                        Navigator.push(
+                            context,
+                            MaterialPageRoute(
+                                builder: (context) => PostListScreen(
+                                    pageTitle: '${category[index]['text']}')));
+                      },
+                      text: '${category[index]['text']}');
                 },
-                text: '최신글'),
-            const SizedBox(
-              height: 20,
-            ),
-            BoardButton(
-                onPressed: () {
-                  context.push('/free');
-                },
-                text: '자유 게시판'),
-            const SizedBox(
-              height: 20,
-            ),
-            BoardButton(
-                onPressed: () {
-                  context.push('/economy');
-                },
-                text: '경제/금융'),
-            const SizedBox(
-              height: 20,
-            ),
-            BoardButton(
-                onPressed: () {
-                  context.push('/lifestyle');
-                },
-                text: '생활'),
-            const SizedBox(
-              height: 20,
-            ),
-            BoardButton(
-                onPressed: () {
-                  context.push('/vision');
-                },
-                text: '진로'),
-            const RootTab()
+              ),
+            )
           ],
         ),
       ),
