+++ carea/lib/app/modules/community/view/post_list.dart
@@ -1,0 +1,134 @@
+import 'package:carea/app/common/const/app_colors.dart';
+import 'package:carea/app/common/layout/default_layout.dart';
+import 'package:carea/app/modules/community/view/add_post.dart';
+import 'package:carea/app/modules/community/view/post_detail.dart';
+import 'package:flutter/material.dart';
+import 'package:carea/app/common/component/category.dart';
+
+class Post {
+  final String title;
+  final String content;
+  final String timestamp;
+  final String author;
+
+  Post(this.title, this.content, this.timestamp, this.author);
+}
+
+class PostListScreen extends StatefulWidget {
+  final String pageTitle;
+
+  const PostListScreen({Key? key, required this.pageTitle}) : super(key: key);
+
+  @override
+  State<PostListScreen> createState() => _PostListScreenState();
+}
+
+class _PostListScreenState extends State<PostListScreen> {
+  final List<Post> posts = List.generate(
+    15,
+    (index) => Post(
+      "게시글 ${index + 1}",
+      "게시글 내용입니다.",
+      '00:00',
+      '젤리',
+    ),
+  );
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        leading: IconButton(
+            onPressed: () {
+              Navigator.pop(context);
+            },
+            icon: const Icon(Icons.keyboard_arrow_left)),
+        title: Center(
+          child: Text(widget.pageTitle),
+        ),
+        actions: [
+          IconButton(
+            onPressed: () {},
+            icon: const Icon(Icons.search),
+          )
+        ],
+        backgroundColor: AppColors.greenPrimaryColor,
+        elevation: 0,
+      ),
+      body: DefaultLayout(
+        child: ListView.builder(
+          padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 15),
+          itemCount: posts.length,
+          itemBuilder: (BuildContext context, int index) {
+            return InkWell(
+                onTap: () {
+                  Navigator.push(
+                    context,
+                    MaterialPageRoute(builder: (context) => const PostDetail()),
+                  );
+                },
+                child: Column(
+                  children: [
+                    ListTile(
+                      title: Text(
+                        posts[index].title,
+                        style: const TextStyle(
+                          fontSize: 18,
+                          fontWeight: FontWeight.bold,
+                        ),
+                      ),
+                      subtitle: Column(
+                        crossAxisAlignment: CrossAxisAlignment.start,
+                        children: [
+                          const SizedBox(height: 10),
+                          Text(
+                            posts[index].content,
+                            style: const TextStyle(fontSize: 16),
+                            maxLines: 2,
+                            overflow: TextOverflow.ellipsis,
+                          ),
+                          const SizedBox(height: 10),
+                          Row(
+                            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                            children: [
+                              Text(
+                                posts[index].timestamp,
+                                style: const TextStyle(
+                                  fontSize: 14,
+                                  color: Colors.grey,
+                                ),
+                              ),
+                              Text(
+                                posts[index].author,
+                                style: const TextStyle(
+                                  fontSize: 14,
+                                  color: Colors.grey,
+                                ),
+                              ),
+                            ],
+                          ),
+                        ],
+                      ),
+                    ),
+                    const Divider(
+                      color: Colors.grey,
+                      thickness: 1,
+                    ),
+                  ],
+                ));
+          },
+        ),
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: () {
+          Navigator.push(
+            context,
+            MaterialPageRoute(builder: (context) => const AddPost()),
+          );
+        },
+        backgroundColor: AppColors.greenPrimaryColor,
+        child: const Icon(Icons.add),
+      ),
+    );
+  }
+}
